Complete Chapter 9, add index counter.
As an extra credit assignment to the chapter,
after essential components were completed,
it calls to create a counter that increments
every time a user calls the store#index action.

This counter is stored in the session.

The counter is reset every time the user adds
an item to his cart.

The store/index template only displays the counter
when the quantity of visits is more than 5.